feat: Implement user incident management and permuta functionalities
- Added MisIncidencias component for users to view their assigned incidents with filtering options.
- Created MisPermutas component to manage proposed permutas, including acceptance and denial functionalities.
- Developed Permutas component to display interesting permutas for users to accept.
- Introduced PermutasAceptadas component to show accepted permutas and allow users to generate drafts.
- Implemented ReportarIncidencia component for users to report incidents with file upload capability.
- Added SeleccionarEstudio component for users to select their study program.
- Created SeleccionarGrupos component for users to select groups for their subjects.
- Developed SelectorAsignatura component for users to select subjects based on their study program.
- Implemented SolicitarPermuta component for users to request permutas.
- Added SolicitudesPermuta component to display user requests and allow filtering by status.
