+++ src/components/comun/RedirectHome.jsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "../hooks/useAuth";
+import { useAuth } from "../../hooks/useAuth";
 
 export default function RedirectHome() {
   const { isAuthenticated, user } = useAuth();
