+++ src/components/comun/footer.jsx
@@ -1,5 +1,5 @@
 import "../styles/footer-style.css";
-import { yearValue } from "../lib/generadorFechas"
+import { yearValue } from "../../lib/generadorFechas"
 
 export default function Footer() {
   return (
