+++ src/components/comun/home.jsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from "react";
 import "../styles/home-style.css";
-import { obtenerNotificaciones } from "../services/notificacion.js";
-import { formatearFecha } from "../lib/formateadorFechas.js";
-import { logError } from "../lib/logger.js";
+import { obtenerNotificaciones } from "../../services/notificacion.js";
+import { formatearFecha } from "../../lib/formateadorFechas.js";
+import { logError } from "../../lib/logger.js";
 
 export default function Home() {
   const [notificaciones, setNotificaciones] = useState([]);
