+++ src/components/comun/noRegistrado.jsx
