+++ src/components/usuario/permutas.jsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from "react";
 import "../styles/permutas-style.css";
-import { obtenerPermutasInteresantes, aceptarPermutaSolicitudesPermuta } from "../services/permuta.js";
+import { obtenerPermutasInteresantes, aceptarPermutaSolicitudesPermuta } from "../../services/permuta.js";
 import { useNavigate } from "react-router-dom";
-import { logError } from "../lib/logger.js";
+import { logError } from "../../lib/logger.js";
 import { toast } from "react-toastify";
 
 export default function Permutas() {
